Bug 934779: Fix signed/unsigned comparisons in TaskbarWindowPreview.cpp and nsClipboard.cpp. r=jimm
UltraBlame original commit: cd1a07426d4c80c65fe44a7795740fd8786ef391

+++ widget/windows/nsClipboard.cpp
@@ -728,8 +728,10 @@ nsClipboard :: FindPlatformHTML ( IDataObject* inDataObject, UINT inIndex, void*
   }
 
   
+  
+  
   if (!endOfData || startOfData >= endOfData || 
-      endOfData > *outDataLen) {
+      static_cast<uint32_t>(endOfData) > *outDataLen) {
     return false;
   }
